FIX order in __init__.py due to init()
When installing the module syndicom_vollzug in an empty
database, the file models/vollzug_declaration_check.py

The file models/vollzug_declaration_check.py contains an
init() method, that executes when the module is installed.
This method calls _query(), that refers to the table
syndicom_vollzug_contact_type. This is the table for the
model syndicom.vollzug.contact.type, defined in the file
models/vollzug_contact_type, that is loaded after in the
__init__.py. As a result, when the init() is executed,
the table is not yet created, thus an exception for
psycopg2.errors.UndefinedTable is raised, blocking the
installation.

One solution is to move the file down in the __init__.py.