[t154397] CLA field for relations; dead-code; licence
When filtering records in two wizards according to the
relation used, a system's parameter was used to define
the relation to use in the filtering condition.

The problems were two: 1) Using a system's parameter is
a bad idea, and 2) There was the need to filter by more
than one relation type, thus the SQL queries had to be
changed anyway.

To deal with the need of filtering by more than one
relation type, and to get rid of the system's parameter,
a new flag has been added to the relation type, named
'Use for CLA Imputed?'. Views have been updated accordingly.

Also, dead-code has been removed, that was found while
searching for the old system's parameter variable. Since
the method was not used anywhere, it was deleted.

The licence for the module has been set as AGPL-3 given
that it depends on the module for partner_multi_relation,
licenced as AGPL-3.