[t154404][FIX] Apply duration_correction to the record duration
This change also adds duration_correction to the api.depends fields, to
run the computation right at the moment the user has modified its value.

Don't check for record.declaration_id.id since this is not going to be
a true value most of the time, since user edition while the main record
has not yet being save returns a temporal id that is evaluated as False,
even if the record exists already in the front-end and can be used
already in the computation.