[t156241][FIX] Remove unused fields
The fields defined as a default value a non-existing ID for
a res.partner. This made the automated tests to crash. This
hardcoding of IDs is a bad practice that we could solve by
simply removing the definition of the fields, since they are
not being used anywhere (other than in the views).