[FIX] Include archived relations when retrieving memberships
Some of the relations were automatically archived by Odoo, since the
date range were not covering the whole current year. When records are
archived, Odoo by default won't return them when querying the database,
if we don't request this expressly.

In reality, we should focus on the filtering we have already in place,
where we look at the date_start and date_end. If the relation is
archived at the present day but applied in its date range we have to use
it.